integrated albums into track viewer
