+++ src/graphql/fragment/DetailedTrack.fragment.graphql
@@ -11,6 +11,7 @@ fragment DetailedTrack on Track {
   album {
     images
     id 
+    name
   }
   artists {
     name
